+++ script/cli/check_worker.lua
@@ -17,6 +17,100 @@ require 'vm'
 
 local export = {}
 
+local colors
+
+if not os.getenv('NO_COLOR') then
+    colors = {
+        red = '\27[31m',
+        green = '\27[32m',
+        yellow = '\27[33m',
+        blue = '\27[34m',
+        magenta = '\27[35m',
+        white = '\27[37m',
+        grey = '\27[90m',
+        reset = '\27[0m'
+    }
+else
+    colors = {
+        red = '',
+        green = '',
+        yellow = '',
+        blue = '',
+        magenta = '',
+        white = '',
+        grey = '',
+        reset = ''
+    }
+end
+
+--- @type table<DiagnosticSeverity, string>
+local severity_colors = {
+    Error = colors.red,
+    Warning = colors.yellow,
+    Information = colors.white,
+    Hint = colors.white,
+}
+
+local severity_str = {} --- @type table<integer,DiagnosticSeverity>
+for k, v in pairs(define.DiagnosticSeverity) do
+    severity_str[v] = k
+end
+
+local pwd
+
+---@param path string
+---@return string
+local function relpath(path)
+    if not pwd then
+        pwd = furi.decode(furi.encode(fs.current_path():string()))
+    end
+    if pwd and path:sub(1, #pwd) == pwd then
+        path = path:sub(#pwd + 2)
+    end
+    return path
+end
+
+local function report_pretty(uri, diags)
+    local path = relpath(furi.decode(uri))
+
+    local lines = {} --- @type string[]
+    pcall(function()
+        for line in io.lines(path) do
+            table.insert(lines, line)
+        end
+    end)
+
+    for _, d in ipairs(diags) do
+        local rstart = d.range.start
+        local rend = d.range['end']
+        local severity = severity_str[d.severity]
+        print(
+            ('%s%s:%s:%s%s [%s%s%s] %s %s(%s)%s'):format(
+                colors.blue,
+                path,
+                rstart.line + 1, -- Use 1-based indexing
+                rstart.character + 1, -- Use 1-based indexing
+                colors.reset,
+                severity_colors[severity],
+                severity,
+                colors.reset,
+                d.message,
+                colors.magenta,
+                d.code,
+                colors.reset
+            )
+        )
+        if #lines > 0 then
+            io.write('    ', lines[rstart.line + 1], '\n')
+            io.write('    ', colors.grey, (' '):rep(rstart.character), '^')
+            if rstart.line == rend.line then
+                io.write(('^'):rep(rend.character - rstart.character - 1))
+            end
+            io.write(colors.reset, '\n')
+        end
+    end
+end
+
 local function clear_line()
     -- Write out empty space to ensure that the previous lien is cleared.
     io.write('\x0D', (' '):rep(80), '\x0D')
@@ -86,6 +180,7 @@ function export.runCLI()
 
     local numThreads = tonumber(NUM_THREADS or 1)
     local threadId = tonumber(THREAD_ID or 1)
+    local quiet = QUIET or numThreads > 1
 
     if type(CHECK_WORKER) ~= 'string' then
         print(lang.script('CLI_CHECK_ERROR_TYPE', type(CHECK_WORKER)))
@@ -127,9 +222,13 @@ function export.runCLI()
 
         client:register('textDocument/publishDiagnostics', function (params)
             results[params.uri] = params.diagnostics
+            if not QUIET and (CHECK_FORMAT == nil or CHECK_FORMAT == 'pretty') then
+                clear_line()
+                report_pretty(params.uri, params.diagnostics)
+            end
         end)
 
-        if not QUIET then
+        if not quiet then
             io.write(lang.script('CLI_CHECK_INITING'))
         end
 
@@ -153,15 +252,15 @@ function export.runCLI()
                 diag.doDiagnostic(uri, true)
                 -- Print regularly but always print the last entry to ensure
                 -- that logs written to files don't look incomplete.
-                if not QUIET and (os.clock() - lastClock > 0.2 or i == #uris) then
+                if not quiet and (os.clock() - lastClock > 0.2 or i == #uris) then
                     lastClock = os.clock()
                     client:update()
                     report_progress(i, max, results)
                 end
             end
         end
-        if not QUIET then
-            io.write('\x0D')
+        if not quiet then
+            clear_line()
         end
     end)
 
@@ -173,16 +272,21 @@ function export.runCLI()
         end
     end
 
-    local outpath = CHECK_OUT_PATH or LOGPATH .. '/check.json'
+    local outpath = nil
 
-    -- Always write result, even if it's empty to make sure no one accidentally looks at an old output after a successful run.
-    util.saveFile(outpath, jsonb.beautify(results))
+    if CHECK_FORMAT == 'json' or CHECK_OUT_PATH then
+        outpath = CHECK_OUT_PATH or LOGPATH .. '/check.json'
+        -- Always write result, even if it's empty to make sure no one accidentally looks at an old output after a successful run.
+        util.saveFile(outpath, jsonb.beautify(results))
+    end
 
-    if not QUIET then
+    if not quiet then
         if count == 0 then
             print(lang.script('CLI_CHECK_SUCCESS'))
+        elseif outpath then
+            print(lang.script('CLI_CHECK_RESULTS_OUTPATH', count, outpath))
         else
-            print(lang.script('CLI_CHECK_RESULTS', count, outpath))
+            print(lang.script('CLI_CHECK_RESULTS_PRETTY', count))
         end
     end
     return count == 0 and 0 or 1